apiapp/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
SQLAlchemy==2.0.43
python-jose==3.5.0
python-multipart==0.0.20
psycopg2-binary==2.9.10
cryptography==45.0.6
requests==2.32.5
python-dotenv==1.1.1
azure-storage-blob==12.26.0 
azure-identity==1.24.0
pandas==2.3.2
python-decouple==3.8
alembic==1.16.5
PyJWT==2.10.1
clickhouse_connect==0.8.18
pydantic_settings==2.10.1
dnspython==2.7.0
email-validator==2.3.0
google-generativeai==0.8.5
tabulate==0.9.0
ollama==0.5.4
openai==2.3.0
pwdlib==0.3.0
argon2-cffi==25.1.0
